feat: added charity validation schemas

+++ server/src/modules/charity/charity.controllers.ts
@@ -5,6 +5,10 @@ import { CharityService } from './charity.service.js'
 import winstonLogger from '../../utils/winston/winstonLogger.js'
 import { CacheService } from '../../services/cache/cache.service.js'
 import { cacheTtlOneDay } from '../../configs/cacheTtl/index.js'
+import {
+  createCharitySchema,
+  updateCharitySchema,
+} from '../../models/Charity.js'
 
 const charityService = new CharityService()
 const cacheService = new CacheService()
@@ -16,15 +20,21 @@ export async function createCharity(
   const { body } = req
 
   try {
-    const charity = await charityService.createCharity(body)
+    const validation = await createCharitySchema.safeParseAsync(body)
 
-    await cacheService.set({
-      key: charity.id,
-      value: charity,
-      expiry: cacheTtlOneDay,
-    })
+    if (validation.success) {
+      const charity = await charityService.createCharity(validation.data)
 
-    res.status(StatusCodes.CREATED).json(charity)
+      await cacheService.set({
+        key: charity.id,
+        value: charity,
+        expiry: cacheTtlOneDay,
+      })
+
+      res.status(StatusCodes.CREATED).json(charity)
+    } else {
+      res.status(StatusCodes.BAD_REQUEST).json(validation.error)
+    }
   } catch (error) {
     winstonLogger.error(error)
 
@@ -83,37 +93,45 @@ export async function getCharityById(
   } = req
 
   try {
-    const cachedCharity = await cacheService.get(id)
+    if (id !== undefined) {
+      const cachedCharity = await cacheService.get(id)
 
-    if (cachedCharity !== null) {
-      res.status(StatusCodes.OK).json({
-        reason: ReasonPhrases.OK,
-        message: 'Successfully found charity from database.',
-        data: cachedCharity,
+      if (cachedCharity !== null) {
+        res.status(StatusCodes.OK).json({
+          reason: ReasonPhrases.OK,
+          message: 'Successfully found charity from database.',
+          data: cachedCharity,
+        })
+      }
+
+      const charity: Charity | null = await charityService.findCharity({ id })
+
+      if (charity !== null) {
+        await cacheService.set({
+          key: charity.id,
+          value: charity,
+          expiry: cacheTtlOneDay,
+        })
+
+        res.status(StatusCodes.OK).json({
+          reason: ReasonPhrases.OK,
+          message: 'Successfully found charity from database.',
+          data: charity,
+        })
+      }
+
+      res.status(StatusCodes.NOT_FOUND).json({
+        reason: ReasonPhrases.NOT_FOUND,
+        message: 'Charity not found.',
+        data: null,
+      })
+    } else {
+      res.status(StatusCodes.BAD_REQUEST).json({
+        reason: ReasonPhrases.BAD_REQUEST,
+        message: 'Charity not found.',
+        data: null,
       })
     }
-
-    const charity: Charity | null = await charityService.findCharity({ id })
-
-    if (charity !== null) {
-      await cacheService.set({
-        key: charity.id,
-        value: charity,
-        expiry: cacheTtlOneDay,
-      })
-
-      res.status(StatusCodes.OK).json({
-        reason: ReasonPhrases.OK,
-        message: 'Successfully found charity from database.',
-        data: charity,
-      })
-    }
-
-    res.status(StatusCodes.NOT_FOUND).json({
-      reason: ReasonPhrases.NOT_FOUND,
-      message: 'Charity not found.',
-      data: null,
-    })
   } catch (error) {
     winstonLogger.error(error)
 
@@ -133,22 +151,32 @@ export async function updateCharityById(
   } = req
 
   try {
-    const updatedCharity = await charityService.updateCharity({
-      data: body,
-      where: { id },
-    })
+    const validation = await updateCharitySchema.safeParseAsync(body)
 
-    await cacheService.set({
-      key: updatedCharity.id,
-      value: updatedCharity,
-      expiry: cacheTtlOneDay,
-    })
+    if (validation.success) {
+      const updatedCharity = await charityService.updateCharity({
+        data: validation.data,
+        where: { id },
+      })
 
-    res.status(StatusCodes.OK).json({
-      reason: ReasonPhrases.OK,
-      message: 'Successfully updated charity.',
-      data: updatedCharity,
-    })
+      await cacheService.set({
+        key: updatedCharity.id,
+        value: updatedCharity,
+        expiry: cacheTtlOneDay,
+      })
+
+      res.status(StatusCodes.OK).json({
+        reason: ReasonPhrases.OK,
+        message: 'Successfully updated charity.',
+        data: updatedCharity,
+      })
+    } else {
+      res.status(StatusCodes.BAD_REQUEST).json({
+        reason: ReasonPhrases.BAD_REQUEST,
+        message: 'Successfully updated charity.',
+        data: null,
+      })
+    }
   } catch (error) {
     winstonLogger.error(error)
 
@@ -167,17 +195,25 @@ export async function deleteCharityById(
   } = req
 
   try {
-    await charityService.deleteCharity({
-      id,
-    })
+    if (id !== undefined) {
+      await charityService.deleteCharity({
+        id,
+      })
 
-    await cacheService.delete(id)
+      await cacheService.delete(id)
 
-    res.status(StatusCodes.OK).json({
-      reason: ReasonPhrases.OK,
-      message: 'Successfully deleted charity.',
-      data: null,
-    })
+      res.status(StatusCodes.OK).json({
+        reason: ReasonPhrases.OK,
+        message: 'Successfully deleted charity.',
+        data: null,
+      })
+    } else {
+      res.status(StatusCodes.BAD_REQUEST).json({
+        reason: ReasonPhrases.BAD_REQUEST,
+        message: 'Successfully deleted charity.',
+        data: null,
+      })
+    }
   } catch (error) {
     winstonLogger.error(error)
